requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mongoengine==0.29.1
pydantic-settings==2.4.0
python-dotenv==1.0.1
redis==5.0.8
rq==1.16.2
rq-scheduler==0.13.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
certifi==2024.8.30
